Fix bug that will try to remove all server data

+++ Envoy.blade.php
@@ -50,9 +50,9 @@
 
 @task('clean_old_releases')
     # This will list our releases by modification time and delete all but the 3 most recent.
-    purging=$(ls -dt {{ $release_dir }}/* | tail -n +3);
+    purging=$(ls -dt {{ $releases_dir }}/* | tail -n +3);
 
-    if [ "$purging" != "" ]; then
+    if [ $releases_dir -ne "" ] && [ $purging -ne "" ]; then
         echo Purging old releases: $purging;
         rm -rf $purging;
     else
